# Make minmax faster for Float32/64