@Seelengrab 2021-07-26
The speedup for `Flaot32/64` is nice, but I'm not sure why allocating more for `BigFloat` is more correct? I'd have expected a non-allocating version to be more in line with what `min` and `max` are supposed to do, namely returning the min/max of its arguments (not a copy thereof).