@N5N3 2021-07-27
Since `BigFloat` is introduced in `Base.MPFR`, maybe such change should be done in a seperated PR?
`Base.Math` only offers the generic version of `AbstractFloat`'s `min/max/minmax`, and make them fast for `IEEEFloat`.