@Seelengrab 2021-07-27
If the change in allocations for `BigFloat` wasn't intended and happened incidentally, I'd be in favor of just doing it in this PR since it touches the same (user facing) API, touching a different file seems fine with me. I think it'd just be a method specialization for `BigFloat` with a non-allocating version, no? Does `MPFR` expose some `min`/`max` API we can use directly without having to implement it ourselves?