@N5N3 2021-07-27
I'm not familiar with `libmpfr`, but Julia now use the 4 args api (2 inputs + output + rounding setting) to calculate the min/max of `BigFloat`, and the following code return false, as the default prec is 256.
```julia
julia> a = BigFloat(1//3,300); b = BigFloat(1//5,300);
julia> min(a, b) == b
false
```
It seems to be a better example to motivate us to the change?