@N5N3 2022-01-10
I think we'd better pick this PR up. @oscardssmith @tkf 
List for triage:
1. Do we really need allocation in `min`/`max` for `BigFloat`.
2. Should we treat `copysign(NaN, -1) < NaN` in `min/max` for `AbstractFloat`.