@tkf 2022-01-18
I consider `min`/`max` to be defined with respect to total orderings [^1] `isless` and `Base.isgreater` since this is a crucial property for parallelism. The result happens to coincide with IEEE 754 for floats and the "poisoning" semantics of `missing`. Since `min`/`max`/`minimum`/`maximum`/`extrema`/etc. provide the "fist match" semantics, I think it makes sense to reuse the input, even for BigFloat NaNs.

[^1]: Or maybe more precisely "partial orderings that are total on the many of the subsets of values people tend to care"