@N5N3 2022-01-19
`BigFloat` is a little bit complex because of it's "variable" precison.
Should we do promotion in `min(x::BigFloat, y::BigFloat)` if they have different precision?