@tkf 2022-01-20
Hmm.... can the promotion of the precision be added to `promote` on `BigFloat`s? Though I guess that's too breaking...