@N5N3 2022-01-20
If we do that, I guess all other math calls should be modified:
```julia
julia> b = BigFloat(1.0,3000) + BigFloat(1.0,200)
2.0
julia> b.prec
256
```
Maybe we can avoid allocation in `min`/`max` only when all inputs has the default precision?