@tkf 2022-01-20
Actually, I have no idea what users would expect about the output precision of `BigFloat` operations. I'm no expert so I'll shut up :)