@oscardssmith 2022-01-20
For `BigFloat` there's no reason to promote (it would just be slower and not more useful). The last thing to check is that this does correct things with `-0.0`. Is this code path sufficiently tested?