@N5N3 2022-01-21
I extend the current `min/max/minmax` test for `Float64` to all float types in `Base`. (5dd01a23bada5afc5fd08afbfea6861e939c800b)
Local test shows no problem.

`_extrema_rf` related optimization is also added in this PR, since it's `min/max` related.