@oscardssmith 2022-06-22
How big is the regression in speed on M1 for this? If it's not major, I'd be in favor of merging.