@JeffreySarnoff 2022-06-22
Each of the functions `min(x,y)`, `max(x,y)`, `minmax(x,y)` use this:
   `anynan = isnan(x) | isnan(y)`

consider
```
for (T,U) in ((:Float64, :UInt64). (:Float32, :UInt32), (:Float16, :UInt16))
  @eval anynan(x::$T, y::$T) =
    isnan( reinterpret($T, 
      ( reinterpret($U, x) | reinterpret($U, y)) 
    )
end
```