@mikmoore 2022-06-22
> Each of the functions `min(x,y)`, `max(x,y)`, `minmax(x,y)` use this: `anynan = isnan(x) | isnan(y)`
> 
> consider
> 
> ```
> for (T,U) in ((:Float64, :UInt64). (:Float32, :UInt32), (:Float16, :UInt16))
>   @eval anynan(x::$T, y::$T) =
>     isnan( reinterpret($T, 
>       ( reinterpret($U, x) | reinterpret($U, y)) 
>     )
> end
> ```

This will erroneously report true for many input combinations such as `anynan(Inf,1.5)`. But that aside, the `isnan(x)|isnan(y)` evaluation can actually be evaluated in a single instruction on x86 (a check that `x` and `y` compare as unordered), so nothing else could be appreciably faster. I can't speak for other architectures.