@JeffreySarnoff 2022-06-22
Ok, there is a bitop fix for my error, but nevermind.
Does Julia or the LLVM use that single instruction version?