@mikmoore 2022-06-22
> Does Julia or the LLVM use that single instruction version?

The `code_native` for this `min` on `Float64`:
```
        vsubsd  %xmm1, %xmm0, %xmm2
        vmovq   %xmm2, %rax
        vmovapd %xmm1, %xmm3
        testq   %rax, %rax
        jns     .LBB0_2
# %bb.1:                                # %top
        vmovapd %xmm0, %xmm3
.LBB0_2:                                # %top
        vcmpordsd       %xmm1, %xmm0, %xmm0
        vblendvpd       %xmm0, %xmm3, %xmm2, %xmm0
        retq
```
The `vcmpordsd` and `vblendvpd` correspond to the `anynan` determination and `ifelse` selection, respectively.